# cmd/docs/swagger.yaml
basePath: /api/v1
definitions:
  attendance.Attendance:
    properties:
      check_in:
        type: string
      check_out:
        type: string
      date:
        type: string
      id:
        type: integer
      status:
        $ref: '#/definitions/attendance.AttendanceState'
      worker:
        $ref: '#/definitions/workers.Worker'
      worker_id:
        type: integer
    type: object
  attendance.AttendanceState:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - CheckedIn
    - CheckedOut
    - Expired
...
  common.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  departments.Department:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  dtos.CreateWorkerDto:
    properties:
      department_id:
        type: integer
      email:
        type: string
      hire_date:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - name
    type: object
...
  dtos.UpdateWorkerDto:
    properties:
      department_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - id
    - name
    type: object
  pagination.PaginationResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  report.WorkerAttendanceReport:
    properties:
      attendanceRate:
        type: number
      daysAbsent:
        type: integer
      daysPresent:
        type: integer
      department:
        type: string
      hoursWorked:
        type: number
      workerId:
        type: integer
      workerName:
        type: string
    type: object
  workers.Worker:
    properties:
      department:
        $ref: '#/definitions/departments.Department'
      department_id:
        type: integer
      email:
        type: string
      hire_date:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      salary:
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: Simple ERP API
  title: User API
  version: "1.0"
paths:
...
  /attendances/worker/{worker_id}:
    get:
      description: Get all attendance records for a specific worker
      parameters:
      - description: Worker ID
        in: path
        name: worker_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/attendance.Attendance'
            type: array
      security:
      - BearerAuth: []
      summary: Get attendance records by worker ID
      tags:
      - attendances
...
  /reports/workers/attendance:
    get:
      description: Get aggregated attendance data for workers within a date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      - description: Filter by department ID
        in: query
        name: department_id
        type: integer
      - description: Filter by worker ID
        in: query
        name: worker_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/report.WorkerAttendanceReport'
            type: array
      security:
      - BearerAuth: []
      summary: Get worker attendance report
      tags:
      - reports
  /workers:
    get:
      description: Get workers with optional pagination query parameters
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 10)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pagination.PaginationResponse'
      security:
      - BearerAuth: []
      summary: List workers (paginated)
      tags:
      - workers
    post:
      consumes:
      - application/json
      description: Create a new worker
      parameters:
      - description: Worker payload
        in: body
        name: worker
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateWorkerDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.CreateWorkerDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIError'
      security:
      - BearerAuth: []
      summary: Create a worker
      tags:
      - workers
    put:
      consumes:
      - application/json
      description: Update an existing worker
      parameters:
      - description: Worker payload
        in: body
        name: worker
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateWorkerDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.UpdateWorkerDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.APIError'
      security:
      - BearerAuth: []
      summary: Update a worker
      tags:
      - workers
  /workers/{id}:
    delete:
      description: Delete a worker by ID
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete a worker
      tags:
      - workers
    get:
      description: Get a worker by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/workers.Worker'
      security:
      - BearerAuth: []
      summary: Get a worker
      tags:
      - workers
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer <JWT_TOKEN>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"